+++ src/main/java/com/leanplatform/MentorshipPlatform/controllers/ScheduleController.java
@@ -1,5 +1,6 @@
 package com.leanplatform.MentorshipPlatform.controllers;
 
+import com.leanplatform.MentorshipPlatform.dto.ScheduleController.CreateScheduleRequest;
 import com.leanplatform.MentorshipPlatform.dto.ScheduleController.CreateScheduleResponse;
 import com.leanplatform.MentorshipPlatform.dto.ScheduleController.DeleteSchedule;
 import com.leanplatform.MentorshipPlatform.dto.ScheduleController.GetAllScheduleResponse;
@@ -19,9 +20,19 @@ public class ScheduleController {
     @Autowired
     ScheduleService scheduleService;
     @PostMapping("/createSchedule")
-    public ResponseEntity<CreateScheduleResponse>createASchedule(@RequestParam (name = "userId") UUID userId, @RequestBody String name){
+    public ResponseEntity<CreateScheduleResponse>createASchedule(@RequestParam (name = "userId") UUID userId, @RequestBody CreateScheduleRequest createScheduleRequest){
+        if(userId==null || createScheduleRequest==null || createScheduleRequest.getName()==null){
+            return new ResponseEntity<>(new CreateScheduleResponse
+                    (
+                            "0",
+                            "Invalid Request",
+                            null
+                    ), HttpStatus.BAD_REQUEST);
+
+
+        }
         try {
-            return scheduleService.createSchedules( userId);
+            return scheduleService.createSchedules( userId,createScheduleRequest);
         } catch (Exception e) {
             return new ResponseEntity<>(new CreateScheduleResponse
                     (
@@ -34,6 +45,14 @@ public class ScheduleController {
     }
     @GetMapping("/getAllSchedule")
     public ResponseEntity<GetAllScheduleResponse>getAllSchedule(@RequestParam(name = "userId") UUID userId){
+        if(userId==null){
+            return new ResponseEntity<>(new GetAllScheduleResponse
+                    (
+                            "0",
+                            "Invalid request",
+                            null
+                    ), HttpStatus.BAD_REQUEST);
+        }
         try {
             return scheduleService.getSchedules( userId);
         } catch (Exception e) {
@@ -47,6 +66,17 @@ public class ScheduleController {
     }
     @GetMapping("/getSchdeule/{scheduleId}")
     public ResponseEntity<CreateScheduleResponse>getSchedule(@PathVariable UUID scheduleId , @RequestParam(name = "userId") UUID userId) {
+        if(userId==null || scheduleId==null){
+            return new ResponseEntity<>(new CreateScheduleResponse
+                    (
+                            "0",
+                            "Invalid Request",
+                            null
+                    ), HttpStatus.BAD_REQUEST);
+
+
+        }
+
         try {
             return scheduleService.getSchdeule(scheduleId, userId);
         } catch (Exception e) {
@@ -61,6 +91,16 @@ public class ScheduleController {
     }
     @DeleteMapping("/deleteScheduleId{scheduleId}")
     public ResponseEntity<DeleteSchedule>deleteSchedule(@PathVariable UUID scheduleId,@RequestParam(name = "userId") UUID userId){
+        if(userId==null || scheduleId==null){
+            return new ResponseEntity<>(new DeleteSchedule
+                    (
+                            "0",
+                            "Invalid Request",
+                            null
+                    ), HttpStatus.BAD_REQUEST);
+
+
+        }
         try {
             return scheduleService.deleteSchedule(scheduleId,userId);
         } catch (Exception e) {
